tweaks to get coverage working
