+++ Shared.EventStore.Tests/AggregateRepositoryManagerTests.cs
@@ -5,6 +5,7 @@ using Aggregate;
 using DomainDrivenDesign.EventSourcing;
 using EventStore;
 using Moq;
+using Shared.EventStore.Tests.TestObjects;
 using Shouldly;
 using Xunit;
