+++ Shared.EventStore.Tests/AggregateRepositoryTests.cs
@@ -10,6 +10,7 @@ using DomainDrivenDesign.EventSourcing;
 using EventStore;
 using global::EventStore.Client;
 using Moq;
+using Shared.EventStore.Tests.TestObjects;
 using Shouldly;
 using Xunit;
