+++ Shared.EventStore.Tests/ServiceCollectionExtensionsTests.cs
@@ -1,0 +1,10 @@
+﻿namespace Shared.EventStore.Tests;
+
+using System.Threading.Tasks;
+using Xunit;
+
+public class ServiceCollectionExtensionsTests{
+    [Fact]
+    public async Task ServiceCollectionExtensions_AddEventStoreProjectionManagerClient(){
+    }
+}
