+++ Shared.EventStore.Tests/SubscriptionRepositoryTests.cs
@@ -4,6 +4,7 @@ namespace Shared.EventStore.Tests
     using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
+    using Shared.EventStore.Tests.TestObjects;
     using Shouldly;
     using SubscriptionWorker;
     using Xunit;
