+++ Shared.EventStore.Tests/SubscriptionWorkerHelperTests.cs
@@ -5,6 +5,7 @@ namespace Shared.EventStore.Tests
     using System.Threading.Tasks;
     using EventHandling;
     using Moq;
+    using Shared.EventStore.Tests.TestObjects;
     using Shouldly;
     using SubscriptionWorker;
     using Xunit;
