+++ Shared.EventStore.Tests/TestObjects/TestData.cs
@@ -1,4 +1,4 @@
-namespace Shared.EventStore.Tests
+namespace Shared.EventStore.Tests.TestObjects
 {
     using System;
     using System.Collections.Generic;
@@ -18,14 +18,14 @@ namespace Shared.EventStore.Tests
         public static Guid AggregateId = Guid.Parse("103B335B-540A-4985-BB80-FD9B2BABF866");
         public static Guid EventId = Guid.Parse("C9416757-582C-4F67-A320-80FE1E937045");
 
-        public static String EstateName = "Test Estate 1";
+        public static string EstateName = "Test Estate 1";
 
-        public static EventRecord CreateEventRecord<T>(T domainEvent, String streamId, Boolean addToMap = true) where T : DomainEvent
+        public static EventRecord CreateEventRecord<T>(T domainEvent, string streamId, bool addToMap = true) where T : DomainEvent
         {
-            Byte[] eventData = Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(domainEvent));
-            Byte[] customEventMetaData = Encoding.UTF8.GetBytes(String.Empty);
+            byte[] eventData = Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(domainEvent));
+            byte[] customEventMetaData = Encoding.UTF8.GetBytes(string.Empty);
 
-            Dictionary<String, String> metaData = new Dictionary<String, String>();
+            Dictionary<string, string> metaData = new Dictionary<string, string>();
             metaData.Add("type", domainEvent.GetType().FullName);
             metaData.Add("created", "1000000");
             metaData.Add("content-type", "application-json");
@@ -48,31 +48,31 @@ namespace Shared.EventStore.Tests
             List<PersistentSubscriptionInfo> list = new List<PersistentSubscriptionInfo>();
 
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "DemoEstate",
-                         GroupName = "Estate Management"
-                     });
+            {
+                StreamName = "DemoEstate",
+                GroupName = "Estate Management"
+            });
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "$et-EstateCreatedEvent",
-                         GroupName = "Migrations"
-                     });
+            {
+                StreamName = "$et-EstateCreatedEvent",
+                GroupName = "Migrations"
+            });
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "FileProcessorSubscriptionStream_DemoEstate",
-                         GroupName = "File Processor"
-                     });
+            {
+                StreamName = "FileProcessorSubscriptionStream_DemoEstate",
+                GroupName = "File Processor"
+            });
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "TransactionProcessorSubscriptionStream_DemoEstate",
-                         GroupName = "Transaction Processor"
-                     });
+            {
+                StreamName = "TransactionProcessorSubscriptionStream_DemoEstate",
+                GroupName = "Transaction Processor"
+            });
 
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "$ce-MerchantBalanceArchive",
-                         GroupName = "Ordered"
-                     });
+            {
+                StreamName = "$ce-MerchantBalanceArchive",
+                GroupName = "Ordered"
+            });
 
             return list;
         }
@@ -82,37 +82,37 @@ namespace Shared.EventStore.Tests
             List<PersistentSubscriptionInfo> list = new List<PersistentSubscriptionInfo>();
 
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "DemoEstate",
-                         GroupName = "Estate Management"
-                     });
+            {
+                StreamName = "DemoEstate",
+                GroupName = "Estate Management"
+            });
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "$et-EstateCreatedEvent",
-                         GroupName = "Migrations"
-                     });
+            {
+                StreamName = "$et-EstateCreatedEvent",
+                GroupName = "Migrations"
+            });
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "FileProcessorSubscriptionStream_DemoEstate",
-                         GroupName = "File Processor"
-                     });
+            {
+                StreamName = "FileProcessorSubscriptionStream_DemoEstate",
+                GroupName = "File Processor"
+            });
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "TransactionProcessorSubscriptionStream_DemoEstate",
-                         GroupName = "Transaction Processor"
-                     });
+            {
+                StreamName = "TransactionProcessorSubscriptionStream_DemoEstate",
+                GroupName = "Transaction Processor"
+            });
 
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "$ce-MerchantBalanceArchive",
-                         GroupName = "Ordered"
-                     });
+            {
+                StreamName = "$ce-MerchantBalanceArchive",
+                GroupName = "Ordered"
+            });
 
             list.Add(new PersistentSubscriptionInfo
-                     {
-                         StreamName = "DemoEstate2",
-                         GroupName = "Estate Management"
-                     });
+            {
+                StreamName = "DemoEstate2",
+                GroupName = "Estate Management"
+            });
 
             return list;
         }
