+++ Shared.EventStore.Tests/TestObjects/TestDomainEventHandler.cs
@@ -1,0 +1,17 @@
+namespace Shared.EventStore.Tests.TestObjects;
+
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using DomainDrivenDesign.EventSourcing;
+using EventHandling;
+
+public class TestDomainEventHandler : IDomainEventHandler
+{
+    public List<IDomainEvent> DomainEvents = new();
+
+    public async Task Handle(IDomainEvent domainEvent, CancellationToken cancellationToken)
+    {
+        DomainEvents.Add(domainEvent);
+    }
+}
