+++ Shared.EventStore.Tests/TestObjects/TestServiceProvider.cs
@@ -1,0 +1,11 @@
+﻿namespace Shared.EventStore.Tests.TestObjects;
+
+using System;
+
+public class TestServiceProvider : IServiceProvider
+{
+    public object GetService(Type serviceType)
+    {
+        return null;
+    }
+}
