+++ Shared.EventStore.Tests/TestObjects/TestState.cs
@@ -1,0 +1,9 @@
+﻿namespace Shared.EventStore.Tests.TestObjects;
+
+using System;
+using ProjectionEngine;
+
+public record TestState : State
+{
+    public string Name { get; set; }
+}
