--- Shared.EventStore.Tests/certs/ca/ca.crt
@@ -1,23 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDzTCCArWgAwIBAgIRAJvALktWBf/BpB4TMM1D91UwDQYJKoZIhvcNAQELBQAw
-YjELMAkGA1UEBhMCVUsxGDAWBgNVBAoTD0V2ZW50IFN0b3JlIEx0ZDE5MDcGA1UE
-AxMwRXZlbnRTdG9yZURCIENBIDliYzAyZTRiNTYwNWZmYzFhNDFlMTMzMGNkNDNm
-NzU1MB4XDTIzMDgwNDE2MzcxNloXDTI4MDgwNDE2MzcxNlowYjELMAkGA1UEBhMC
-VUsxGDAWBgNVBAoTD0V2ZW50IFN0b3JlIEx0ZDE5MDcGA1UEAxMwRXZlbnRTdG9y
-ZURCIENBIDliYzAyZTRiNTYwNWZmYzFhNDFlMTMzMGNkNDNmNzU1MIIBIjANBgkq
-hkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAy7kBNjcPrg2C/m+5VvL3fIUw2GJU8tKN
-9bOebSoa0WzOVXYDaBnmXLEn74jTfl2NSJsoi4YCfB87xOY40QnelZyd7804XZee
-hOY2hVCr1nwB9LGtRI0ZQ5Rt0Kew7mS0c8IAcoktOpYz9wR13uEP1l8lebXC/amm
-47CzRHJsA4uuIwPNTGZJoCVjSg0pu3wNEbMyCVSdwbUPhSzk/HHDace1rRAZ87fT
-aRgT1vInjpt3WUlJmqsWpiZ6ZvPHPL3FUac4nS37Q/IbuoHPcoHhP1exF8VtZktN
-UEztTxmishcukaDDvyIztiwFQfh5B8uYYJJjNUnQKQXDWfTaQK58aQIDAQABo34w
-fDAOBgNVHQ8BAf8EBAMCAQYwEgYDVR0TAQH/BAgwBgEB/wIBAjApBgNVHQ4EIgQg
-sMxaF0ihJl7xXyPiuP+MvqbUENNk6uKn5oqLq2P2DT4wKwYDVR0jBCQwIoAgsMxa
-F0ihJl7xXyPiuP+MvqbUENNk6uKn5oqLq2P2DT4wDQYJKoZIhvcNAQELBQADggEB
-AKv+JoBMV+sm2jA7ZGYNTOhjiewu0UJ0g95R+OF6vAoyUx9m8yoOg1zjEuoEG7Ap
-xuLTm0QObKhtIs6rMH/LyGTznHFpKVRr8qkAV4b1K6wtODmhTEOwnS0y0OPQGnaz
-O/vOL+4gFEp5t4WHotvuTcikbf7lF9OTW1W5VoBF1qbWV2znf5BO62FsIO/lDYeh
-ZFkx7wg0y1SFTXxjG6chAk6+ZSU6mtsH4ceRJK1ioofY95xAYMjSH4MK5z5saZ9C
-skX5UdtgIFwbTOgmVZQ3H9xnLLoe7aFkXyLmdbmzPRYoUTbaR37JFMYQ2Z7JATSk
-HhuSLxZ4kucBG4GmGoEIidA=
------END CERTIFICATE-----
