--- Shared.EventStore.Tests/certs/ca/ca.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEAy7kBNjcPrg2C/m+5VvL3fIUw2GJU8tKN9bOebSoa0WzOVXYD
-aBnmXLEn74jTfl2NSJsoi4YCfB87xOY40QnelZyd7804XZeehOY2hVCr1nwB9LGt
-RI0ZQ5Rt0Kew7mS0c8IAcoktOpYz9wR13uEP1l8lebXC/amm47CzRHJsA4uuIwPN
-TGZJoCVjSg0pu3wNEbMyCVSdwbUPhSzk/HHDace1rRAZ87fTaRgT1vInjpt3WUlJ
-mqsWpiZ6ZvPHPL3FUac4nS37Q/IbuoHPcoHhP1exF8VtZktNUEztTxmishcukaDD
-vyIztiwFQfh5B8uYYJJjNUnQKQXDWfTaQK58aQIDAQABAoIBAQCiAnBaQxNvIUOp
-QAQuh24VUn5ztk9EA/ntCXjV64QEQ/kasY0n/mDaGfBeZ3t5pMA0m191BY94UgZg
-IY3Rf33JkRFwoM1QcMEZg1KcVnkxStDrXVkjiOoseh+SrajhDKkM3Z0peLnFO6n7
-qUprgqProSLWVZvJyl7z95O4vR30+nZAHZgdIRZWtTsH/iZCBd9ZVIhK+dO3xoQS
-OENbwSmEv0nL4qdW9IDuyRChRIa8xYPQiePst6L8ToJFFQugGutEH+4S6eLeQl8N
-/2iohDGwb3owG1oBAlAhhITMS81XAKx4UNhgSywrwwKHe0sBXuGuA1t0e9pXlHPO
-mVi1pyrhAoGBAPBVjpzDuxrxKtZaWKjuVtV8w8GPIBxoRE1lO5R/dk9W+lBN3zWx
-MPoznqLPpYNoBDHy6lkVvM7KHH82Q7XzOxAdlgIqpAKV2JxmatvnIt77L7wBYWus
-DtHH6pf30iQHrHq/28skhZxNHfutyRjECybENBQ69R0k8qSTyqtHM/s1AoGBANkA
-g11CIyTmYy0KJSbxD47/l/MtKnE18hmaRwyf+OccfpibelGpCh2Oli3e93zq7eon
-1Wc3zkktG6RFSrEOiQcUhFZo/YfSpenHLIV1ZpoIL309B7iJtvN2mYg29zTGxznj
-/vzIh9dW54RBvnzWyTieQmHDCSZUyJv21hq+Bm7lAoGBAM7i18FbPNYUhSvA2lGR
-2m4s2Nbmvl215BY77u4GS3JBDXnTjzJJNkAr74N3/KMCB5p34gSOEdpxYdagf+lB
-WLamwVtcGZOgKrd38sUaU4pf8NeG2LY3hwpPbkMjsmIRuQhx4GMfLWU/fXAKSfJn
-0LGudKZ1mktcYwQJtY173eOBAoGAZcBR2IbMh97nnhTlus7J5tl+XG5POmp7LPok
-Wo0gXu4DhndQ8b17msRjMvsbvR1Dlco7zHxkY/ecwHglP8R7CDDcpM/yIZ/qYCYe
-gC2+gHdaHTViy2r3uE1iTHbJj5utiz2XMNyupZCAWomhUlQ8YioDImgbgDe9ITSq
-zDtSNOkCgYEA6qKnbcgA7mE7Bi3X5ab4CsFbIvD8GuD1ngrrxhz0CMjfEdMufpoc
-keQD5Wvaqt2tAQSduh4tnGT2me3LIw2lexnQXwv6JwWR3TmXinZEOWQuPjHjti3M
-DBoM8jLTr/Pm9heEnc5G1klccY51t486PPCuQjo2FUw0AnviqoESDLU=
------END RSA PRIVATE KEY-----
