--- Shared.EventStore.Tests/certs/node1/node.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDszCCApugAwIBAgIRANn8h5xWhI1/AZ05RJ00lZUwDQYJKoZIhvcNAQELBQAw
-YjELMAkGA1UEBhMCVUsxGDAWBgNVBAoTD0V2ZW50IFN0b3JlIEx0ZDE5MDcGA1UE
-AxMwRXZlbnRTdG9yZURCIENBIDliYzAyZTRiNTYwNWZmYzFhNDFlMTMzMGNkNDNm
-NzU1MB4XDTIzMDgwNDE2NDE0N1oXDTI0MDgwNDE2NDE0N1owHDEaMBgGA1UEAxMR
-ZXZlbnRzdG9yZWRiLW5vZGUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIB
-AQC9FvRLc6hSEjLIijLUSEKvFiKxu6Vj+Lqs9jQOdipezvYtToR1+4kso+l2KaM2
-Iki0Z2wqgn3MUnCT5yj4YBGwrilv7SFCTp00SfoDeQ/hArI4YY2xfr18kn8tn2MB
-6Oq6aIwESMwxq1uvQhF9L23vXCJsCXx7UP6xXPEQ9qerg5ANnNvNUC5kJbuIOEGN
-Bjfry44RBTj0K//u98ifX6BiqAR0gkmtU1IaJmPzI6ttm6Jf5IBPgG+klCzEO7JH
-ioTksB4GYshVdgzk6dzbSHaEY1YgcbyXDkvJohYQ2YMAycS2iOo21BiUXvaQv4ep
-l1QeuoLNip7i8LN3nTHPxfopAgMBAAGjgakwgaYwDgYDVR0PAQH/BAQDAgWgMB0G
-A1UdJQQWMBQGCCsGAQUFBwMCBggrBgEFBQcDATAMBgNVHRMBAf8EAjAAMCkGA1Ud
-DgQiBCCTkPPERnrsbq5jhjBaAcZosZptAs4UinaHJrSe1ffn2TArBgNVHSMEJDAi
-gCCwzFoXSKEmXvFfI+K4/4y+ptQQ02Tq4qfmiourY/YNPjAPBgNVHREECDAGhwR/
-AAABMA0GCSqGSIb3DQEBCwUAA4IBAQCEHSoK9ZFhQql2y5kriGG96ovLD3+luwBb
-H9qJNSR6Alxx8V/HSR9u1/68sF4Ia67eQEAyOPc4rQNRL6/0IAxFItpaNmGgyrDU
-EYT/FikPnyZEvqnDdae5KQ+TPTMZPeBfjcsVJCtrNGL0eNWgCpgSGZ1skZa3m7z9
-ZdmU6DBtreXsdxrKhDUiSzA7QIds4Bikcvu/EO/RR9wn8v3qcbbe5rboWg5AtDEF
-EhuyIG1fAiWIWxmLmYNTI5noOW1O2twq8IV32Ewal1BGmmTntoQcGDShCVkq9B91
-NTYf/0OxUrokTZ2+LZlpR34C+SJ10pxlvf9e+1lniGWqKIcKQ859
------END CERTIFICATE-----
