--- Shared.EventStore.Tests/certs/node1/node.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEowIBAAKCAQEAvRb0S3OoUhIyyIoy1EhCrxYisbulY/i6rPY0DnYqXs72LU6E
-dfuJLKPpdimjNiJItGdsKoJ9zFJwk+co+GARsK4pb+0hQk6dNEn6A3kP4QKyOGGN
-sX69fJJ/LZ9jAejqumiMBEjMMatbr0IRfS9t71wibAl8e1D+sVzxEPanq4OQDZzb
-zVAuZCW7iDhBjQY368uOEQU49Cv/7vfIn1+gYqgEdIJJrVNSGiZj8yOrbZuiX+SA
-T4BvpJQsxDuyR4qE5LAeBmLIVXYM5Onc20h2hGNWIHG8lw5LyaIWENmDAMnEtojq
-NtQYlF72kL+HqZdUHrqCzYqe4vCzd50xz8X6KQIDAQABAoIBAQCw2boso4SnFew3
-UcIOd3Me7HYQgy9B3E1EHZU8fVpNiAYJA+mxq0t12+wrnu/OXqxfRmeNLrTXl7dB
-La9M9hU1iP6qdz9nE73GPPY/r0ehrHLhdqQegSpaS231TqA8FgsLw2JJU4jfX6SF
-uyY+TlDk0OhE4W5yq6M6YXH56qeRKgR0Ju4RdRGh0GTJ1THRyo1S14iHpDhitW1A
-IPR0Yv9RQdSFOOcjzkdjqk7WWWzrrui1sT4/BSG9Pq7R8eAJCDmU2VECQKyURLIf
-z0MNyXLHDdYGKXlupI6JL2RDltR56mUv7vcEtgRUZURF0sGh1LI7Emr0V92YZ4gA
-2tiouoP9AoGBANMHNzncmNagsGS4vL7mHzkkIA/QNPwgvjtX6AllaMvtoEqVVoIw
-q2NbSmBiBbjalaPZ8LrXHHstLvd9cF2oFRHOl0AQDAb+KAGlXfkVtxa9f/MQw/7u
-wJTcWm1R9JfwYktxyGieFEMda7vteW+4o8wSBp43yYwNPjbnzFRug+VTAoGBAOVi
-3ihjqoMdxot/EmntUBJocQucbkGZlsGqCNuCJ0uncLD+BF0nY/itc/QpPdewiADr
-Bjgl7Xa8GO6XJt8mc4esDj4YWHBniJ9Jg6IhDoNQFrDoUBek8hj7/4/ocR6h9wU3
-4z0INDMxRUuBos5nGG01RD3mozL8CHIxrGoCaJcTAoGAa0OfrqmShoXnI0tDpUUT
-Wdgi8+bBB+ccys1EBIKnYgDHGSIPdt2GQFqG07b/cZapuT9uPYpZz8KHvQTluLVX
-AFYiS03czzEg5LVJrvJ6vfRUtRgwh8ST3xaiYC7oRNLLI84SMBsp8XNi74Ml1sPj
-fjcrZgC1UZrppL7t/XyvTkECgYBJQO6TVWjBoIOkTrtRqqCUkFfxjk5DksnrJOLZ
-PtJPvbQmQaihGxA4/SN3iKggNlbofCWrTqlr7pvDre+S+dORMygFOn8bvmSKAc8Q
-up/I3Y2xhy70rg9GXUt66I8fajfZqcRe3DzXUFfByi903MWXMpiNcBcUFBqaZTSb
-UhOWkwKBgDmQh6LeRQFVTQ/ihaXv18/zDSzdRxMN5vPGWi/8ch/Rn1GDhJDetv+A
-Xnguo3Q2zPAnGbEUa6iw6Lb2T7+581dqMsT0jH+z2OQwKmwn/yCDkwZCl+Vrd2ft
-lwXVkJiZG7EcmKVIOyWtAF6qx+UcbOeRggfB/zEx1UXxmcZBUnMT
------END RSA PRIVATE KEY-----
